Populated owners info.

+++ webinar-08-authorization/repositories/cats.js
@@ -1,27 +1,33 @@
 const Cat = require('../model/cat');
 
-const getAllCats = async () => {
-  const results = await Cat.find();
+const getAllCats = async userId => {
+  const results = await Cat.find({ owner: userId }).populate({
+    path: 'owner',
+    select: 'name email gender -_id',
+  });
   return results;
 };
 
-const getCatById = async id => {
-  const result = await Cat.findOne({ _id: id });
+const getCatById = async (userId, id) => {
+  const result = await Cat.findOne({ _id: id, owner: userId }).populate({
+    path: 'owner',
+    select: 'name email gender -_id',
+  });
   return result;
 };
 
-const removeCat = async id => {
-  const result = await Cat.findOneAndRemove({ _id: id });
+const removeCat = async (userId, id) => {
+  const result = await Cat.findOneAndRemove({ _id: id, owner: userId });
   return result;
 };
 
-const addCat = async body => {
-  const result = await Cat.create(body);
+const addCat = async (userId, body) => {
+  const result = await Cat.create({ ...body, owner: userId });
   return result;
 };
 
-const updateCat = async (id, body) => {
-  const result = await Cat.findOneAndUpdate({ _id: id }, { ...body }, { new: true });
+const updateCat = async (userId, id, body) => {
+  const result = await Cat.findOneAndUpdate({ _id: id, owner: userId }, { ...body }, { new: true });
   return result;
 };
